Bit more component abstraction in the header nav
