+++ apps/web/src/components/navigation/header/links.ts
@@ -1,0 +1,36 @@
+const links = [
+  {
+    slug: '/',
+    text: 'Home',
+  },
+  {
+    slug: '/about',
+    text: 'About',
+  },
+  {
+    slug: '/portfolio',
+    text: 'Portfolio',
+  },
+  {
+    slug: '/music',
+    text: 'Music',
+  },
+  {
+    slug: '/events',
+    text: 'Events',
+  },
+  {
+    slug: '/talks',
+    text: 'Speaking',
+  },
+  {
+    slug: '/blog',
+    text: 'Blog',
+  },
+  {
+    slug: '/contact',
+    text: 'Contact',
+  },
+];
+
+export { links }
